+++ src/Interface.h
@@ -6,7 +6,7 @@ class Interface
 {
 	public:
 		//todo in theory if this is the only function that gets called through the backend this could be the constructor
-		Interface(const int& curLivPop, const int& curWorkPop, const int& turn, const float& devotion);
+		Interface(const int& curLivPop, const int& curWorkPop, const int& turn, const int& devotion);
 
 
 
@@ -17,12 +17,12 @@ class Interface
 		void updateCLP(const int& curLivPop);
 		void updateCWP(const int& curWorkPop);
 		void updateT(const int& turn);
-		void updateDev(const float& devotion);
+		void updateDev(const int& devotion);
 	private:
 		int m_clP = 0;
 		int m_cwP = 0;
 		int m_t = 0;
-		float m_dev = 0.0f;
+		int m_dev = 0;
 		
 		char m_i;
